requirements.txt
groq
python-dotenv
flake8
gitpython